+++ indexer/f95zone.py
@@ -1,0 +1,88 @@
+import collections
+import contextlib
+import datetime as dt
+import logging
+import os
+import sys
+
+import aiohttp
+import aiolimiter
+
+XENFORO_RATELIMIT = aiolimiter.AsyncLimiter(max_rate=6, time_period=2)
+TIMEOUT = 30
+LOGIN_ERROR_MESSAGES = (
+    b'<a href="/login/" data-xf-click="overlay">Log in or register now.</a>',
+    b"<title>Log in | F95zone</title>",
+)
+RATELIMIT_ERROR_MESSAGES = (
+    b"<title>Error 429</title>",
+    b"<title>DDOS-GUARD</title>",
+)
+TEMP_ERROR_MESSAGES = (
+    b"<title>502 Bad Gateway</title>",
+    b"<!-- Too many connections -->",
+    b"<p>Automated backups are currently executing. During this time, the site will be unavailable</p>",
+)
+
+logger = logging.getLogger(__name__)
+session: aiohttp.ClientSession = None
+cookies: dict = None
+
+THREAD_URL = "https://f95zone.to/threads/{thread}"
+MASKED_URL = "https://f95zone.to/masked/"
+
+IndexerError = collections.namedtuple(
+    "IndexerError",
+    (
+        "error_flag",
+        "retry_delay",
+    ),
+)
+
+ERROR_SESSION_LOGGED_OUT = IndexerError(
+    "SESSION_LOGGED_OUT", dt.timedelta(hours=2).total_seconds()
+)
+ERROR_XENFORO_RATELIMIT = IndexerError(
+    "XENFORO_RATELIMIT", dt.timedelta(minutes=15).total_seconds()
+)
+ERROR_F95ZONE_UNAVAILABLE = IndexerError(
+    "F95ZONE_UNAVAILABLE", dt.timedelta(minutes=15).total_seconds()
+)
+
+
+@contextlib.asynccontextmanager
+async def lifespan(version: str):
+    global session, cookies
+    session = aiohttp.ClientSession(cookie_jar=aiohttp.DummyCookieJar())
+    session.headers["User-Agent"] = (
+        f"F95Indexer/{version} "
+        f"Python/{sys.version.split(' ')[0]} "
+        f"aiohttp/{aiohttp.__version__}"
+    )
+    cookies = {
+        "xf_user": os.environ.get("COOKIE_XF_USER"),
+    }
+
+    try:
+        yield
+    finally:
+
+        await session.close()
+        session = None
+
+
+def check_error(res: bytes) -> IndexError | None:
+    if any((msg in res) for msg in LOGIN_ERROR_MESSAGES):
+        logger.error("Logged out of F95zone")
+        # TODO: maybe auto login, but xf_user cookie should be enough for a long time
+        return ERROR_SESSION_LOGGED_OUT
+
+    if any((msg in res) for msg in RATELIMIT_ERROR_MESSAGES):
+        logger.error("Hit F95zone ratelimit")
+        return ERROR_XENFORO_RATELIMIT
+
+    if any((msg in res) for msg in TEMP_ERROR_MESSAGES):
+        logger.warning("F95zone temporarily unreachable")
+        return ERROR_F95ZONE_UNAVAILABLE
+
+    return None
